# docs/openapi/core.yaml
openapi: 3.1.0
info:
  title: Tilecast API
  version: 0.10.0
  license:
    name: AGPL-3.0-only
  description: >-
    Management authentication accepts either an enrolled dashboard session
    cookie or a bearer user grant (an OAuth access token or a personal
    access token), and both produce the same user principal with the live
    role and screen scope. Grants additionally carry a scope ceiling of
    read, write, or admin (admin implies write implies read): reads take
    read, mutations take write, and user, token, grant, backup, and
    security administration take admin. CSRF is cookie-only: unsafe
    browser requests require X-CSRF-Token, bearer requests never send it.
    Browser security ceremonies (MFA, passkeys, logout, OAuth approval)
    stay session-only.
paths:
...
  /api/v1/settings:
    get:
      operationId: getSettings
      security: [{ sessionCookie: [] }, { userBearer: [] }]
      description: >-
        Requires an authenticated user with the read scope. Reads the
        typed organization settings document with its registry and
        revision.
      responses:
        "200": { description: Settings document and registry }
        "401": { description: Dashboard authentication required }
    patch:
      operationId: updateSettings
      security: [{ sessionCookie: [] }, { userBearer: [] }]
      description: >-
        Requires the Owner or Administrator role with the write scope.
        Cookie requests additionally require the X-CSRF-Token header;
        bearer requests never send it. Applies only the given keys
        against the expected revision; a stale revision fails with 409 and
        changes nothing.
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdate" }
      responses:
        "200": { description: Updated settings }
        "401": { description: Dashboard authentication required }
        "403": { description: Owner or Administrator and CSRF token required }
        "409": { description: Settings revision conflict }
        "422": { description: "Unknown, invalid, or hard-limit setting" }
  /api/v1/me/preferences:
    get:
      operationId: getPreferences
      description: >-
        Requires an enrolled dashboard session. Reads the calling user's
        own preferences.
      responses:
        "200": { description: Preference document }
        "401": { description: Dashboard authentication required }
    patch:
      operationId: updatePreferences
      description: >-
        Requires an enrolled dashboard session and the X-CSRF-Token
        header. Updates the calling user's own preferences against the
        expected revision.
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdate" }
      responses:
        "200": { description: Updated preferences }
        "401": { description: Dashboard authentication required }
        "403": { description: CSRF token required }
        "409": { description: Revision conflict }
...
  /api/v1/screens/{id}/effective-policy:
    get:
      operationId: getEffectivePolicy
      description: >-
        Requires an enrolled dashboard session. Reports the effective
        player policy for one screen with inheritance sources. A screen
        outside the caller's scope answers 404.
      parameters:
        - { $ref: "#/components/parameters/ResourceID" }
      responses:
        "200": { description: Effective policy }
        "401": { description: Dashboard authentication required }
        "404": { description: Screen unknown or outside caller scope }
...
  /api/v1/system/status:
    get:
      operationId: systemStatus
      description: >-
        Requires the Owner or Administrator role and an enrolled dashboard
        session. Secret-free system diagnostics: migration version,
        database, connectivity, and queue depths.
      responses:
        "200": { description: Secret-free status }
        "401": { description: Dashboard authentication required }
        "403": { description: Owner or Administrator required }
...
components:
  parameters:
...
    ResourceID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    CSRFToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
      description: Cookie-backed browser requests only. Bearer grants never send it.
  securitySchemes:
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: Tilecast device credential
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: Tilecast user grant (tca_ OAuth access token or tcp_ personal access token)
    sessionCookie:
      type: apiKey
      in: cookie
      name: tilecast_session
      description: Enrolled dashboard session cookie
  schemas:
...
    SettingsUpdate:
      type: object
      additionalProperties: false
      required: [revision, values]
      properties:
        revision: { type: integer, format: int64, minimum: 0 }
        values: { type: object }